requirements.txt
Flask
tensorflow
Pillow
numpy
matplotlib
gunicorn
